+++ src/regex.ts
@@ -3,3 +3,7 @@ import { variablePattern } from "./patterns";
 export const variableReg = new RegExp(variablePattern);
 
 export const matchVariableReg = new RegExp(`(\\s+|$\{)(${variablePattern})$`);
+
+export function getFragmentDeclarationReg(name: string) {
+  return new RegExp(`<#${name}\\s+(${variablePattern}).*>`);
+}